+++ tests/Unit/Command/CommandsDebugCommandTest.php
@@ -4,14 +4,19 @@ namespace Tests\OriNette\Console\Unit\Command;
 
 use OriNette\Console\Command\CommandsDebugCommand;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Tester\CommandTester;
+use Tests\OriNette\Console\Doubles\HiddenAndAliasedCommand;
+use Tests\OriNette\Console\Doubles\SimpleCommand;
+use Tests\OriNette\Console\Doubles\UrlPrintingCommand;
 use function array_map;
 use function explode;
 use function implode;
+use function preg_replace;
+use function putenv;
 use function rtrim;
 use const PHP_EOL;
-use const PHP_MAJOR_VERSION;
-use const PHP_MINOR_VERSION;
+use const PHP_VERSION_ID;
 
 final class CommandsDebugCommandTest extends TestCase
 {
@@ -36,32 +41,29 @@ MSG,
 	public function testCommands(): void
 	{
 		$commands = [
-			['service1', true, false],
-			['service2', false, true],
-			['service3', true, true],
+			['service1', UrlPrintingCommand::class, true, false],
+			['service2', HiddenAndAliasedCommand::class, false, true],
+			['service3', SimpleCommand::class, true, true],
+			['service3', Command::class, false, false],
 		];
 		$command = new CommandsDebugCommand($commands);
 		$tester = new CommandTester($command);
 
+		putenv('COLUMNS=80');
 		$code = $tester->execute([]);
 
-		$expected = PHP_MAJOR_VERSION >= 8 && PHP_MINOR_VERSION >= 1 ? <<<'MSG'
-Following commands are missing ❌ either name or description. Check orisai/nette-console documentation about lazy loading to learn how to fix it.
- ---------- ------ -------------
-  Service    Name   Description
-  service1   ✔️     ❌
-  service2   ❌     ✔️
-  service3   ✔️     ✔️
- ---------- ------ -------------
+		if (PHP_VERSION_ID < 8_01_00) {
+			self::markTestSkipped('Printing of X acts weird on PHP < 8.1');
+		}
 
-MSG : <<<'MSG'
+		$expected = <<<'MSG'
 Following commands are missing ❌ either name or description. Check orisai/nette-console documentation about lazy loading to learn how to fix it.
- ---------- ------ -------------
-  Service    Name   Description
-  service1   ✔️     ❌
-  service2   ❌      ✔️
-  service3   ✔️     ✔️
- ---------- ------ -------------
+
+Name Description Service name Service type
+✔️   ❌          service1     Tests\OriNette\Console\Doubles\UrlPrintingCommand
+❌   ✔️          service2     Tests\OriNette\Console\Doubles\HiddenAndAliasedCommand
+✔️   ✔️          service3     Tests\OriNette\Console\Doubles\SimpleCommand
+❌   ❌          service3     Symfony\Component\Console\Command\Command
 
 MSG;
 
@@ -71,7 +73,7 @@ MSG;
 				PHP_EOL,
 				array_map(
 					static fn (string $s): string => rtrim($s),
-					explode(PHP_EOL, $tester->getDisplay()),
+					explode(PHP_EOL, preg_replace('~\R~u', PHP_EOL, $tester->getDisplay())),
 				),
 			),
 		);