+++ tests/Unit/DI/ConsoleExtensionTest.php
@@ -38,10 +38,10 @@ use function dirname;
 use function explode;
 use function implode;
 use function mkdir;
+use function preg_replace;
+use function putenv;
 use function rtrim;
 use const PHP_EOL;
-use const PHP_MAJOR_VERSION;
-use const PHP_MINOR_VERSION;
 use const PHP_VERSION_ID;
 
 final class ConsoleExtensionTest extends TestCase
@@ -583,41 +583,33 @@ MSG);
 
 		$tester = new CommandTester($command);
 
+		putenv('COLUMNS=80');
 		$code = $tester->execute([]);
 
-		$expected = PHP_MAJOR_VERSION >= 8 && PHP_MINOR_VERSION >= 1 ? <<<'MSG'
-Following commands are missing ❌ either name or description. Check orisai/nette-console documentation about lazy loading to learn how to fix it.
- ---------------------------------- ------ -------------
-  Service                            Name   Description
-  command.defaultName                ✔️     ❌
-  command.tagged.name.a              ✔️     ❌
-  command.tagged.name.b              ✔️     ❌
-  command.tagged.name.c              ✔️     ❌
-  command.hiddenAndAliased.negated   ✔️     ❌
-  command.notLazy                    ❌     ✔️
- ---------------------------------- ------ -------------
+		if (PHP_VERSION_ID < 8_01_00) {
+			self::markTestSkipped('Printing of X acts weird on PHP < 8.1');
+		}
 
-MSG : <<<'MSG'
+		$expected = <<<'MSG'
 Following commands are missing ❌ either name or description. Check orisai/nette-console documentation about lazy loading to learn how to fix it.
- ---------------------------------- ------ -------------
-  Service                            Name   Description
-  command.defaultName                ✔️     ❌
-  command.tagged.name.a              ✔️     ❌
-  command.tagged.name.b              ✔️     ❌
-  command.tagged.name.c              ✔️     ❌
-  command.hiddenAndAliased.negated   ✔️     ❌
-  command.notLazy                    ❌      ✔️
- ---------------------------------- ------ -------------
+
+Name Description Service name                     Service type
+✔️   ❌          command.defaultName              Tests\OriNette\Console\Doubles\DefaultNameCommand
+✔️   ❌          command.tagged.name.a            Tests\OriNette\Console\Doubles\SimpleCommand
+✔️   ❌          command.tagged.name.b            Tests\OriNette\Console\Doubles\SimpleCommand
+✔️   ❌          command.tagged.name.c            Tests\OriNette\Console\Doubles\SimpleCommand
+✔️   ❌          command.hiddenAndAliased.negated Tests\OriNette\Console\Doubles\HiddenAndAliasedCommand
+❌   ✔️          command.notLazy                  Symfony\Component\Console\Command\Command
 
 MSG;
 
-		self::assertEquals(
+		self::assertSame(
 			$expected,
 			implode(
 				PHP_EOL,
 				array_map(
 					static fn (string $s): string => rtrim($s),
-					explode(PHP_EOL, $tester->getDisplay()),
+					explode(PHP_EOL, preg_replace('~\R~u', PHP_EOL, $tester->getDisplay())),
 				),
 			),
 		);